requirements.txt
aiobotocore==2.24.2
aiohappyeyeballs==2.6.1
aiohttp==3.12.15
aioitertools==0.12.0
aiosignal==1.4.0
annotated-types==0.7.0
anyio==3.7.1
attrs==25.3.0
botocore==1.40.18
bounded-pool-executor==0.0.3
certifi==2025.8.3
charset-normalizer==3.4.3
click==8.3.0
colorama==0.4.6
coverage==7.10.7
earthaccess==0.15.1
fastapi==0.104.1
frozenlist==1.7.0
fsspec==2025.9.0
h11==0.16.0
httptools==0.6.4
idna==3.10
importlib_resources==6.5.2
iniconfig==2.1.0
jmespath==1.0.1
multidict==6.6.4
multimethod==2.0
packaging==25.0
pluggy==1.6.0
pqdm==0.2.0
propcache==0.3.2
pydantic==2.5.2
pydantic-settings==2.1.0
pydantic_core==2.14.5
pytest==7.4.3
pytest-cov==7.0.0
python-cmr==0.13.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.0
python-multipart==0.0.6
PyYAML==6.0.1
requests==2.31.0
s3fs==2025.9.0
six==1.17.0
sniffio==1.3.1
starlette==0.27.0
tenacity==9.1.2
tinynetrc==1.3.1
tqdm==4.67.1
typing_extensions==4.15.0
urllib3==2.5.0
uvicorn==0.24.0.post1
watchfiles==1.1.0
websockets==15.0.1
wrapt==1.17.3
yarl==1.20.1
